refactor: do not report 4xx to Sentry
Ref. metriport/metriport-internal#909
