+++ packages/api/src/routes/middlewares/usage.ts
@@ -2,7 +2,8 @@ import { NextFunction, Request } from "express";
 import { ApiTypes, reportUsage as reportUsageCmd } from "../../command/usage/report-usage";
 import { Util } from "../../shared/util";
 import { getUserIdFrom } from "../schemas/user-id";
-import { getCxId, isHttpOK } from "../util";
+import { getCxId } from "../util";
+import { isHttpOK } from "../../shared/http";
 
 const log = Util.log("USAGE");
