+++ packages/api/src/routes/util.ts
@@ -1,8 +1,7 @@
 import { NextFunction, Request, Response } from "express";
-import httpStatus from "http-status";
 import { ApiTypes } from "../command/usage/report-usage";
 import BadRequestError from "../errors/bad-request";
-import { analytics, EventTypes } from "../shared/analytics";
+import { EventTypes, analytics } from "../shared/analytics";
 import { Config } from "../shared/config";
 import { capture } from "../shared/notifications";
 
@@ -171,17 +170,3 @@ export const getDateOrFail = (req: Request): string => {
   if (!date) throw new BadRequestError("Missing date query param");
   return date as string;
 };
-
-export function getETag(req: Request): {
-  eTag: string | undefined;
-} {
-  const eTagHeader = req.header("If-Match");
-  const eTagPayload = req.body.eTag;
-  return {
-    eTag: eTagHeader ?? eTagPayload,
-  };
-}
-
-export function isHttpOK(statusCode: number): boolean {
-  return httpStatus[`${statusCode}_CLASS`] === httpStatus.classes.SUCCESSFUL;
-}
